@@ -8,10 +8,9 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0-0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   args: ^2.0.0-nullsafety.0
-  universal_io: ^1.0.2
   boardgame_io:
     path: ../..
