@@ -10,10 +10,8 @@ description: >-
 repository: http://github.com/flarbear/boardgame_io
 
 environment:
-  sdk: '>=2.9.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  universal_io: ^1.0.2
-  http: ^0.12.2
   socket_io_client: ^1.0.0-nullsafety.0
   logging: ^1.0.0-nullsafety.0
